Add New functions to initialize each layer's struct and refactor the directory structure
