--- adapter/handler/handler.go
@@ -1,5 +1,0 @@
-package handler
-
-type VocabHandler struct {
-	Usecase VocabUsecase
-}
