--- adapter/handler/response/error.go
@@ -1,5 +1,0 @@
-package response
-
-type ErrorRes struct {
-	Message string `json:"message"`
-}
