+++ adapter/handler/vocabulary/add.go
@@ -1,13 +1,13 @@
-package handler
+package vocabulary
 
 import (
 	"encoding/json"
 	"log/slog"
 	"net/http"
 
-	"github.com/takumi616/go-restapi-sample/adapter/handler/request"
-	"github.com/takumi616/go-restapi-sample/adapter/handler/response"
-	"github.com/takumi616/go-restapi-sample/adapter/handler/transform"
+	"github.com/takumi616/go-restapi-sample/adapter/handler/vocabulary/request"
+	"github.com/takumi616/go-restapi-sample/adapter/handler/vocabulary/response"
+	"github.com/takumi616/go-restapi-sample/adapter/handler/vocabulary/transform"
 )
 
 func (vh *VocabHandler) AddVocabulary(w http.ResponseWriter, r *http.Request) {
