+++ adapter/handler/vocabulary/handler.go
@@ -1,0 +1,9 @@
+package vocabulary
+
+type VocabHandler struct {
+	Usecase VocabUsecase
+}
+
+func New(vocabUsecase VocabUsecase) *VocabHandler {
+	return &VocabHandler{vocabUsecase}
+}
