--- adapter/repository/repository.go
@@ -1,5 +1,0 @@
-package repository
-
-type VocabRepository struct {
-	Persistence VocabPersistence
-}
