+++ adapter/repository/vocabulary/repository.go
@@ -1,0 +1,9 @@
+package vocabulary
+
+type VocabRepository struct {
+	Persistence VocabPersistence
+}
+
+func New(vocabPersistence VocabPersistence) *VocabRepository {
+	return &VocabRepository{vocabPersistence}
+}
