--- infrastructure/database/crud/persistence.go
@@ -1,7 +1,0 @@
-package crud
-
-import "database/sql"
-
-type VocabPersistence struct {
-	DB *sql.DB
-}
