+++ infrastructure/database/vocabulary/model/vocabulary.go
