+++ infrastructure/database/vocabulary/persistence/create.go
@@ -1,4 +1,4 @@
-package crud
+package persistence
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"log/slog"
 
 	"github.com/takumi616/go-restapi-sample/entity"
-	"github.com/takumi616/go-restapi-sample/infrastructure/database/transform"
+	"github.com/takumi616/go-restapi-sample/infrastructure/database/vocabulary/transform"
 )
 
 func (vp *VocabPersistence) Create(ctx context.Context, vocabulary *entity.Vocabulary) (int64, error) {
