+++ infrastructure/database/vocabulary/persistence/persistence.go
@@ -1,0 +1,11 @@
+package persistence
+
+import "database/sql"
+
+type VocabPersistence struct {
+	DB *sql.DB
+}
+
+func New(db *sql.DB) *VocabPersistence {
+	return &VocabPersistence{db}
+}
