+++ usecase/vocabulary/add.go
@@ -1,4 +1,4 @@
-package usecase
+package vocabulary
 
 import (
 	"context"
@@ -6,10 +6,6 @@ import (
 
 	"github.com/takumi616/go-restapi-sample/entity"
 )
-
-type VocabUsecase struct {
-	Repository VocabRepository
-}
 
 func (vu *VocabUsecase) AddVocabulary(ctx context.Context, vocabulary *entity.Vocabulary) (int64, error) {
 	// Call repository to create a new vocabulary
@@ -19,13 +15,4 @@ func (vu *VocabUsecase) AddVocabulary(ctx context.Context, vocabulary *entity.Vo
 	}
 
 	return rowsAffected, err
-}
-
-func (vu *VocabUsecase) GetVocabularyList(ctx context.Context) ([]*entity.Vocabulary, error) {
-	vocabularyList, err := vu.Repository.FindAll(ctx)
-	if err != nil {
-		slog.ErrorContext(ctx, "found an error returned from adapter layer")
-	}
-
-	return vocabularyList, err
 }
