+++ usecase/vocabulary/repository.go
@@ -1,4 +1,4 @@
-package usecase
+package vocabulary
 
 import (
 	"context"
