+++ usecase/vocabulary/usecase.go
@@ -1,0 +1,9 @@
+package vocabulary
+
+type VocabUsecase struct {
+	Repository VocabRepository
+}
+
+func New(vocabRepository VocabRepository) *VocabUsecase {
+	return &VocabUsecase{vocabRepository}
+}
